Поправил код-стайл, вынес утилиты отдельно от классов
