+++ src/components/search-result-group.js
@@ -2,9 +2,10 @@ import AbstractComponent from './abstract';
 
 class SearchResultGroup extends AbstractComponent {
   getTemplate() {
-    return `<section class="result__group">
-      <div class="result__cards"></div>
-    </section>`.trim();
+    return `
+      <section class="result__group">
+        <div class="result__cards"></div>
+      </section>`.trim();
   }
 }
