+++ src/components/search-result.js
@@ -3,10 +3,9 @@ import AbstractComponent from './abstract';
 class SearchResult extends AbstractComponent {
   getTemplate() {
     return `
-    <section class="result container">
+      <section class="result container">
         <button class="result__back">back</button>
-      </section>
-  `.trim();
+      </section>`.trim();
   }
 }
