+++ src/controllers/search.js
@@ -1,6 +1,5 @@
-import moment from 'moment';
-
 import {render, removeElement, Position} from '../utils/render';
+import {isHasDescription, isHasTag, isHasDate} from '../utils/tasks';
 
 import TaskListController from './task-list';
 
@@ -60,28 +59,6 @@ class SearchController {
 
   _init() {
     this.hide();
-
-    const isHasDescription = (task, value) => task.description.includes(value);
-
-    const isHasTag = (task, value) => {
-      const tagValue = value.replace(`#`, ``);
-      return Array.from(task.tags).find((tag) => tag.includes(tagValue));
-    };
-
-    const isHasDate = (task, value) => {
-      const taskDate = moment(task.dueDate);
-      const values = value.toLowerCase().replace(`d`, ``).split(`.`);
-      let inputDate = ``;
-
-      const isFullValues = () => values.filter((el) => el).length === 3;
-
-      if (isFullValues()) {
-        inputDate = moment(`${values[1]}, ${values[0]}, ${values[2]}`);
-        return taskDate.format(`DD-MM-YYYY`) === inputDate.format(`DD-MM-YYYY`);
-      }
-
-      return false;
-    };
 
     render(this._container, this._searchResult.getElement(), Position.BEFOREEND);
     render(this._searchResult.getElement(), this._searchResultGroup.getElement(), Position.BEFOREEND);
